access: Move TryCompat implementation for BitFlags from compat
This is cleaner and would avoid accessing private fields.
